samples/life-insurance/deploy/requirements.txt
strands-agents[a2a]
bedrock-agentcore
boto3
